adds check to see if that user already exists on the register page, adds style to prevent the footer from covering the bottom of the register form if the screen height is small

+++ app/components/register/register.js
@@ -112,8 +112,15 @@ var Register = React.createClass({
             }
             axios.post('/register', cred)
                 .then(function (response) {
-                    sessionStorage.setItem('do_good_id', response.data);
-                    this.context.router.push('main');
+                    console.log(response.data);
+                    if (response.data === "User already exists") {
+                        this.setState({
+                            error: "User already exists, please select a new username or login"
+                        })
+                    } else {
+                        sessionStorage.setItem('do_good_id', response.data);
+                        this.context.router.push('main');
+                    }
                 }.bind(this))
         }
     },
@@ -135,7 +142,7 @@ var Register = React.createClass({
 
                 <div className="container">
                     <div className="row">
-                        <div className="col-md-12">
+                        <div className="col-md-12 register-margin-bottom">
                             <div className="login-panel panel panel-default">
                                 <div className="panel-heading">
                                     <h3 className="panel-title">Registration</h3>
